ci: new e2e pipeline
